Wireless Network Design
Design a network using a desktop that is connected with a switch which is connected to a router. Further, create a wireless network where different end systems like PC, Laptop, Wireless Tablet, Smartphone, and a printer is connected to a Access point which is connected with the switch. Test the connectivity of the network using packet tracer.